Sizes of the integer types are reconsidered.
git-svn-id: svn+ssh://svn.osdn.jp/svnroot/alchemusica/trunk@117 19e2383c-df9d-4eb5-8775-ae43218f5bcb

+++ Classes/GraphicClientView.h
@@ -100,8 +100,8 @@ enum {
 
 - (int)modifyLocalGraphicTool:(int)originalGraphicTool;
 
-- (void)convertFromPoint:(NSPoint)pt toY:(float *)y andTick:(long *)tick;
-- (NSPoint)convertToPointFromY:(float)y andTick:(long)tick;
+- (void)convertFromPoint:(NSPoint)pt toY:(float *)y andTick:(int32_t *)tick;
+- (NSPoint)convertToPointFromY:(float)y andTick:(int32_t)tick;
 
 - (void)setVisibleRangeMin:(float)min max:(float)max;
 - (void)getVisibleRangeMin:(float *)min max:(float *)max;
